Explicitly exported all symbols. For gcc the macros need to be updated with the visibility attribute. Added VS2010 project files. Updated Window README files to reflect current state of build system. Updated deploy .cmd files appropriately. Did not update VS2003/2005/2008 vcproj files! This must be done before release. Added MSVC intermedidate files to .gitignore BaseExc is no longer derived from std::string (because std::string is not adorned for export from the STL on MSVC) Added operators to BaseExc to preserve string interface Made throwErrnoExc not segfault on windows Fixed miscellaneous precision loss warnings Fixed zlib problems on MSVC (ZLIB_WINAPI must not be specified in EXR source code) Fix MSVC runtime segfault in ScanLineInputFile::readPixels Fix miscellaneous (but not all) signed/unsigned assignment warnings Use !! instead of bool constructor Modify IlmImfTest so that it is runnable on Windows Repair some (but not all) 3 space tabs to 4 spaces Note: testRgba.cpp has no functional change in the diff, only tab replacement

+++ OpenEXR/IlmImf/ImfConvert.cpp
@@ -116,7 +116,7 @@ uintToHalf (unsigned int ui)
     if (ui >  HALF_MAX)
 	return half::posInf();
 
-    return half (ui);
+    return half ((float) ui);
 }
